+++ ingest/metadata_validator.go
@@ -23,7 +23,7 @@ type MetadataValidator struct {
 // to set validator.Profile to a BagItProfile object before calling Run().
 // This is an unfortunate side effect of the need to conform to a common
 // constructor pattern.
-func NewMetadataValidator(context *common.Context, workItemID int, ingestObject *service.IngestObject) *MetadataValidator {
+func NewMetadataValidator(context *common.Context, workItemID int64, ingestObject *service.IngestObject) *MetadataValidator {
 	profileName := ingestObject.BagItProfileFormat()
 	profile, err := loadProfile(context, profileName)
 	if err != nil {
